+++ benker/schemas/__init__.py
@@ -1,7 +1,19 @@
 # coding: utf-8
+"""
+XSD Schemas
+===========
 
+XSD Schemas used by builders.
+"""
+
+#: URI of the CALS schema (fake URL).
 CALS_NS = "https://lib.benker.com/schemas/cals.xsd"
+
+#: Namespace prefix used in XML instances to identify a CALS table.
 CALS_PREFIX = "cals"
 
+#: URI of the Formex schema (fake URL)
 FORMEX_NS = "https://lib.benker.com/schemas/formex.xsd"
+
+#: Namespace prefix used in XML instances to identify a Formex table.
 FORMEX_PREFIX = "fmx"