+++ benker/__init__.py
@@ -1,6 +1,6 @@
 """Easily convert your CALS, HTML, Formex 4, Office Open XML (docx) tables from one format to another."""
 
-__version__ = "0.5.4"
+__version__ = "0.5.5"
 __author__ = "Laurent LAPORTE <laurent.laporte.pro@gmail.com>"
-__date__ = "2021-11-13"
+__date__ = "unreleased"
 __credit__ = "(c) 2018-2021 Laurent LAPORTE"