+++ CHANGELOG.rst
@@ -17,6 +17,12 @@ Fixed
 -----
 
 
+Other
+-----
+
+Change in the documentation: fix a broken link in the documentation.
+
+
 v0.4.4 (2021-11-10)
 ===================