+++ CHANGELOG.rst
@@ -20,7 +20,7 @@ Fixed
 Other
 -----
 
-Change in the documentation: fix a broken link in the documentation.
+Change in the documentation: fix a broken and redirected links in the documentation.
 
 
 v0.4.4 (2021-11-10)
@@ -58,7 +58,7 @@ Other
 -----
 
 Change in the documentation: update the URL of the Formex documentation, use:
-`<https://publications.europa.eu/en/web/eu-vocabularies/formex/>`_.
+`<https://op.europa.eu/en/web/eu-vocabularies/formex/>`_.
 
 
 v0.4.2 (2019-06-06)