+++ CHANGELOG.rst
@@ -7,6 +7,15 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+v0.5.4 (unreleased)
+===================
+
+Bug fix release
+
+Fixed
+-----
+
+
 v0.5.4 (2021-11-13)
 ===================